requirements.txt
python-dotenv
openai
langchain-core==0.3.0
langchain-openai==0.3.0
langchain-anthropic==0.2.0